SAVING AGAIN
Updated the way item data is saved. Before it was saving the object that it was, so if the object did not exist in the scene for the load phase, it would not load since that specific object did not exist. Now we are saving the items with all their info (name, and type) so out save file is now a list of arrays, with each array being the item and the list being the whole inventory (with the item counts for the slots) Next up is the load system.